+++ protocol.py
@@ -42,6 +42,7 @@ MSG_SPEAK = "speak"
 MSG_CANCEL = "cancel"
 MSG_PAUSE_SPEECH = "pause_speech"
 MSG_NVDA_NOT_CONNECTED = "nvda_not_connected"
+MSG_KEY = "key"
 
 
 # Connection type sent in the join message. "master" = we are the